refactor: data module의 DI를 ViewModelComponent에서 SingletonComponent로 변경
